requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
validators==0.22.0
